fix: move `<MakeswiftProvider>` to the root layout + Next 15 compat
